mejora en los modulos de entradas, platos
