+++ app/App.js
@@ -6,9 +6,9 @@ import { StyleSheet, useWindowDimensions } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import CustomDrawerContent from '../components/CustomDrawerContent';
 import Almacenes from '../components/views/Almacenes';
-import Entradas from '../components/views/Entradas';
+import Entradas from '../components/views/Entradas/Entradas';
 import Productos from '../components/views/Insumos/Productos';
-import Platos from '../components/views/Platos';
+import Platos from '../components/views/Platos/Platos';
 import Proveedores from '../components/views/Proveedores';
 import Reportes from '../components/views/Reportes';
 import Salidas from '../components/views/Salidas';
